backend/requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
boto3==1.28.30
botocore==1.31.30
cached-property==1.5.2
click==8.1.7
dictdiffer==0.9.0
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.101.1
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
inflection==0.5.1
iso8601==1.1.0
jmespath==1.0.1
jose==1.0.0
mypy-extensions==1.0.0
passlib==1.7.4
Pillow==10.0.0
psycopg2-binary==2.9.5
pyasn1==0.5.0
pydantic==2.5.3
pydantic_core==2.14.6
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
s3transfer==0.6.2
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.12.3
tortoise-orm==0.20.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
